Reverted to old branch. Moved IDevice outside of interface
