+++ sdks/cpp/common/include/IDevice.h
